requirements.txt
aiogram==3.3.0
httpx==0.27.2
python-dateutil==2.9.0.post0
pytz==2024.1
